requirements.txt
streamlit
openai
python-dotenv
streamlit-echarts
streamlit-extras
boto3
pandas
